fix: move commands to cmd
