+++ cmd/rhmas/completion/completion.go
