+++ cmd/rhmas/kafka/credentials/credentials.go
