+++ cmd/rhmas/kafka/delete/delete.go
