+++ cmd/rhmas/kafka/status/status.go
