+++ cmd/rhmas/logout/logout.go
