+++ cmd/rhmas/root/root.go
@@ -2,15 +2,15 @@ package root
 
 import (
 	"github.com/MakeNowJust/heredoc"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/completion"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/docs"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/kafka"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/login"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/logout"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhmas/completion"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhmas/docs"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhmas/kafka"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhmas/login"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhmas/logout"
 	"github.com/spf13/cobra"
 )
 
-func NewCmdRoot() *cobra.Command {
+func NewRootCommand() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "rhmas <command> <subcommand> [flags]",
 		Short: "RHMAS CLI",
